fix: update balanceAmount to be an integer string and fix unit tests related to balance amount

+++ core_chat_presentation/src/commonMain/kotlin/net/thechance/mena/core_chat/presentation/screen/home/HomeViewModel.kt
@@ -188,7 +188,7 @@ class HomeViewModel(
     }
 
     private fun onGetBalanceAmountSuccess(balanceAmount: Double) {
-        updateState { it.copy(balanceAmount = balanceAmount.toString(), isBalanceLoading = false) }
+        updateState { it.copy(balanceAmount = balanceAmount.toInt().toString(), isBalanceLoading = false) }
     }
 
     private fun onGetBalanceAmountError() {
